move react into peer and dev dependencies
I did some digging and discovered that you can avoid users inadvertently
bundling two versions of React (and all of the other madness that may
come with that) by declaring it as a peer dependency. Because peer deps
aren't installed by default, we also have to add a more specific version
to devDependencies for testing. See this PR for an example:
https://github.com/moroshko/react-autosuggest/pull/26/files