Update README.md
Add an easy link to search through the octicons.

### Why this PR?

So it's easy to click and search through the octicons. I've had to manually navigate it many times myself via Google. Having it here will make it super easy for anyone to what they want to see. Hope it helps - if not feel free to close. chrs