Cache Octicon SVGs for same icon/options combination
This PR upstreams an optimization we've been using
internally to speed up rendering the same octicon
multiple times.

Rendering octicons can be expensive when done often on a page. For
example, rendering the same "check" octicon in a large list of items can
add up on a page, so often we'll render it once and then pass in the
rendered HTML string into the list items.

Co-authored-by: Simon Taranto <srt32@github.com>