update: installation instructions
### Why this change?

* To clarify installation instructions: 

Usage of Rails view helpers to render octicons are discouraged as per #674 - but legacy users might be confused as to how to make it work - given it worked previously. This PR seeks to address any doubts raised. I hope it is useful, if not feel free to close.

rgds
Ben